fix: the translation is incorrect when switching languages for the first time
closes #4138
