+++ packages/locales/src/i18n.ts
@@ -1,8 +1,13 @@
 import type { Locale } from 'vue-i18n';
 
-import type { ImportLocaleFn, SupportedLanguagesType } from './typing';
+import type {
+  ImportLocaleFn,
+  LoadMessageFn,
+  LocaleSetupOptions,
+  SupportedLanguagesType,
+} from './typing';
 
-import { unref } from 'vue';
+import { type App, unref } from 'vue';
 import { createI18n } from 'vue-i18n';
 
 const loadedLanguages = new Set<string>();
@@ -17,6 +22,8 @@ const i18n = createI18n({
 const modules = import.meta.glob('./langs/*.json');
 
 const localesMap = loadLocalesMap(modules);
+
+let loadMessages: LoadMessageFn;
 
 /**
  * Load locale modules
@@ -45,11 +52,28 @@ function setI18nLanguage(locale: Locale) {
   document?.querySelector('html')?.setAttribute('lang', locale);
 }
 
+async function setupI18n(app: App, options: LocaleSetupOptions = {}) {
+  const { defaultLocale = 'zh-CN' } = options;
+  // app可以自行扩展一些第三方库和组件库的国际化
+  loadMessages = options.loadMessages || (async () => ({}));
+  app.use(i18n);
+  await loadLocaleMessages(defaultLocale);
+
+  // 在控制台打印警告
+  i18n.global.setMissingHandler((locale, key) => {
+    if (options.missingWarn && key.includes('.')) {
+      console.warn(
+        `[intlify] Not found '${key}' key in '${locale}' locale messages.`,
+      );
+    }
+  });
+}
+
 /**
  * Load locale messages
  * @param lang
  */
-async function loadI18nMessages(lang: SupportedLanguagesType) {
+async function loadLocaleMessages(lang: SupportedLanguagesType) {
   if (unref(i18n.global.locale) === lang) {
     return setI18nLanguage(lang);
   }
@@ -62,9 +86,12 @@ async function loadI18nMessages(lang: SupportedLanguagesType) {
 
   if (message?.default) {
     i18n.global.setLocaleMessage(lang, message.default);
+
+    const mergeMessage = await loadMessages(lang);
+    i18n.global.mergeLocaleMessage(lang, mergeMessage);
   }
   loadedLanguages.add(lang);
   return setI18nLanguage(lang);
 }
 
-export { i18n, loadI18nMessages, loadLocalesMap, setI18nLanguage };
+export { i18n, loadLocaleMessages, loadLocalesMap, setupI18n };
